requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-cors-headers==3.1.0
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.2
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
PyMySQL==0.9.3
pytz==2019.2
requests==2.22.0
serializers==0.2.4
uritemplate==3.0.0
urllib3==1.25.3
